Add vehicle service and main mission task.
This commit adds a map item initialization script to set up a vehicle
service area from a trigger placed in the editor. It is parameterized by
the base class name of the type of vehicle you would like the area to
service.

This commit also fixes the color of the main target area to be more
visible from the map, adds an icon to the main target area, and creates
a task in the player's journal when the main target has been activated.
When the main target area has been completed to the satisfaction of the
script, the task is also marked completed.

Also adds radio chatter to the HE support option.